requirements.txt
asgiref==3.3.1
Django==3.1.5
django-nested-inline==0.4.2
django-super-inlines==0.1.4
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
